Eliminate response wait time for the user
Waiting for a response for the server is still necessary. However, now in addition to the displayed text, there is also a stored text that is on standby to replace the displayed text. Then, the stored text gets a new value from the server. The wait time for this is not visible to the user.